e2e/gatewayapi: merge selective + rgdisabled into one file
